requirements.txt
feedparser==6.0.8
Jinja2==3.0.3
PyYAML==6.0
